Split blobstorage.h
