+++ ydb/core/tx/columnshard/defs.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "common/blob.h"
 
-#include <ydb/core/base/blobstorage.h>
+#include <ydb/core/base/blobstorage_tablet_types.h>
 #include <ydb/core/base/defs.h>
 #include <ydb/core/base/events.h>
 #include <ydb/core/control/lib/immediate_control_board_impl.h>
